Created About page for the application and added info about us
